refactor(rust): split `node create` command code into separate files for background/foreground modes
